Add files via upload
Reenviando algumas coisas, peguei maioria por segurança mas consegui ajeitar para acessar outras páginas e o h2-console

+++ src/main/java/br/ufms/facom/progweb/services/UsuarioService.java
@@ -1,33 +1,53 @@
 package br.ufms.facom.progweb.services;
 
-import br.ufms.facom.progweb.models.Usuario;
-import br.ufms.facom.progweb.repositories.UsuarioRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import br.ufms.facom.progweb.models.Usuario;
+import br.ufms.facom.progweb.repositories.UsuarioRepository;
+
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UsuarioService {
-    private final UsuarioRepository usuarioRepository;
 
-    public UsuarioService(UsuarioRepository usuarioRepository) {
-        this.usuarioRepository = usuarioRepository;
+    @Autowired
+    private UsuarioRepository usuarioRepository;
+
+    @Autowired
+    private PasswordEncoder encoder;
+
+    public Usuario buscarPorEmail(String email){
+        return usuarioRepository.findByEmail(email).orElse(null);
     }
 
-    public List<Usuario> listarTodos() {
+    public List<Usuario> listarTodos(){
         return usuarioRepository.findAll();
     }
-
-    public Optional<Usuario> buscarPorId(Long id) {
-        return usuarioRepository.findById(id);
+    public void deletar(Long id){
+        if(!usuarioRepository.existsById(id)){
+            throw new RuntimeException("Usuário não encontrado");
+        }
+        usuarioRepository.deleteById(id);
+    }
+    public Usuario buscarPorId(Long id){
+        return usuarioRepository.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
     }
 
-    public Usuario salvar(Usuario usuario) {
+    public Usuario atualizar(Long id, Usuario usuarioAtualizado){
+        Usuario usuario = buscarPorId(id);
+
+        usuario.setNome(usuarioAtualizado.getNome());
+        usuario.setEmail(usuarioAtualizado.getEmail());
         return usuarioRepository.save(usuario);
     }
-
-    public void deletar(Long id) {
-        usuarioRepository.deleteById(id);
+    public Usuario salvar(Usuario usuario){
+        if(buscarPorEmail(usuario.getEmail()) !=null){
+            throw new RuntimeException("Email já cadastrado");
+        }
+        usuario.setSenha(encoder.encode(usuario.getSenha()));
+        usuario.setRole("ROLE_USER");
+        return usuarioRepository.save(usuario);
     }
 }
